Move migrate_html_cache_keys to background service
It's only run from the CLI, and this makes the code more uniform with
the nbstripout service.
